SMTChecker: Fix analysis for selected contracts
Previously, when a contract was selected for analysis, the analysis was
incorrect.
There were two issues. First, the contracts in the same file were
considered as entry points even though they were not selected for
analysis.
Second, the contracts in a different file were mostly ignored, resulting
in unsoundness when an external call was made to such a contract in
trusted mode.

The solution to the above problems is to always create representation of all
contracts (in case they are called from the selected contract), but
create verification targets only for the selected contracts.